+++ jpviz/dot/styling.py
@@ -1,0 +1,44 @@
+GRAPH_STYLING = dict(
+    fontname="Courier",
+    fontsize="12",
+    style="dotted",
+    labeljust="l",
+)
+IN_ARG_STYLING = dict(
+    shape="box",
+    color="green",
+    fontname="Courier",
+    fontsize="10",
+)
+LITERAL_STYLING = dict(
+    shape="box",
+    color="orange",
+    fontname="Courier",
+    fontsize="10",
+)
+OUT_ARG_STYLING = dict(
+    shape="box",
+    color="red",
+    fontname="Courier",
+    fontsize="10",
+)
+VAR_STYLING = dict(
+    shape="box",
+    color="blue",
+    fontname="Courier",
+    fontsize="10",
+)
+PRIMITIVE_STYLING = dict(
+    shape="oval",
+    fontname="Courier",
+    fontsize="10",
+    linestye="dashed",
+)
+COND_NODE_STYLING = dict(
+    shape="box",
+    fontname="Courier",
+    fontsize="10",
+)
+FUNCTION_NODE_STYLING = dict(
+    shape="rectangle", fontname="Courier", fontsize="10", style="dotted"
+)
